+++ ui/src/components/ContentItem.vue
