+++ ui/src/components/icons/IconConfiguration.vue
