+++ ui/src/components/icons/IconVideo.vue
